requirements.txt
Django==4.1.13
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
pymongo[srv]==3.12.3
pytz==2024.1
whitenoise==6.6.0
gunicorn==22.0.0
uvicorn==0.29.0
# daphne==4.1.2 # daphne -b 0.0.0.0 -p 8000 FitnessifyAPI.asgi:application
